Refactor, fixes
