+++ apps/redi-connect/src/pages/app/applications/application-card/ApplicationCard.scss
@@ -18,8 +18,6 @@
 
   // not happy with the solution, time for refactoring together with avatar
   &__avatar {
-    display: flex;
-    align-items: center;
     flex-grow: 0;
 
     .avatar {
@@ -51,9 +49,7 @@
   }
 
   &-dropdown {
-    display: flex;
-    align-items: center;
-    justify-content: center;
+    flex-grow: 0;
 
     .icon {
       transition: $transition;
